SFSIP-112 sofia does not compile w/ gcc 4.3 x86 w/ -Werror
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10877 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/sofia-sip/libsofia-sip-ua/nta/nta.c
@@ -1593,7 +1593,7 @@ int agent_set_params(nta_agent_t *agent, tagi_t *tags)
   if (maxsize > UINT32_MAX) maxsize = UINT32_MAX;
   agent->sa_maxsize = maxsize;
 
-  if (max_proceeding == 0) max_proceeding = SIZE_MAX;
+  if (max_proceeding == 0) max_proceeding = UINT32_MAX;
   agent->sa_max_proceeding = max_proceeding;
 
   if (max_forwards == 0) max_forwards = 70; /* Default value */
